+++ engine/src/point.cc
@@ -5,13 +5,13 @@ Point::Point(): Point(0,0)
 
 }
 
-Point::Point(double a, double b)
+Point::Point(float a, float b)
 {
     x = a;
     y = b;
 }
 
-void Point::Set(double a, double b)
+void Point::Set(float a, float b)
 {
     x = a;
     y = b;
